Moved encoder and decoder blocks into separate files. Minimal refactoring.
